# openapi.yaml
openapi: 3.0.3
info:
  title: Needle Eye API
  version: 0.1.0
  description: |
    Express + TypeScript backend for the Needle Eye boutique/tailoring ERP.
    Stateless REST: every request carries its own bearer token, verified
    independently -- nothing about a client's session is stored server-side
    between requests. `needleye-web` never talks to Supabase directly; every
    auth operation goes through the `/auth/*` endpoints documented here.

    See the repo README for the layered architecture (Controller ->
    Validation -> Service -> Repository -> Mapper) and the Mermaid flow-map
    diagrams -- this document covers the *what* (request/response contracts),
    the README's Flow Map section covers the *how* (which function calls
    which).
  contact:
    name: Needle Eye API
servers:
  - url: http://localhost:4000/api/v1
    description: Local dev
...
security:
  - bearerAuth: []
paths:
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current caller's profile
      description: |
        Verifies the bearer token against Supabase Auth, then loads the
        matching `profiles` row. Every page in needleye-web calls this once
        (server-side, in `app/(app)/layout.tsx`) to know who's logged in and
        what role they have.
      responses:
        "200":
          description: The caller's profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile:
                    $ref: "#/components/schemas/Profile"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/bootstrap:
    post:
      tags: [Auth]
      summary: Create the first Owner/Manager account
      description: |
        Self-disabling: throws 403 the instant one Owner/Manager already
        exists, so it can never be used to self-escalate after go-live.
        Every account after this one is created directly via `POST /users`
        (no email invite -- a password is generated and returned once).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, fullName]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                fullName: { type: string, minLength: 1 }
      responses:
        "201":
          description: Owner/Manager account created
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean, enum: [true] } }
        "403":
          description: An Owner/Manager already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in with email + password
      description: |
        The only place needleye-web sends a raw password. Returns a fresh
        session; if the account turns out to be inactive, the session is
        immediately revoked again server-side before the error is returned --
        a deactivated user never walks away holding a live token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Session established
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate a refresh token for a new session
      description: |
        Called by needleye-web's `proxy.ts` (on every request, if the access
        token is missing/expired) and by `lib/api/client.ts`/`lib/api/server.ts`
        before any API call whose token has expired. Supabase rotates the
        refresh token itself on every use -- the old one stops working the
        moment a new one is issued.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current session
      description: |
        Real server-side revocation (`auth.admin.signOut`), not just "the
        client forgets the token" -- a refresh attempt with the same token
        after this call fails with 401. Best-effort: an already-invalid token
        is not treated as an error.
      responses:
        "204": { description: Session revoked }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/password-reset-request:
    post:
      tags: [Auth]
      summary: Request a password-reset email
      description: |
        Always returns 204, whether or not the email matches an account --
        never confirms account existence. The email's link points at
        `${WEB_APP_URL}/auth/callback?next=/update-password`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses:
        "204": { description: Always -- see description }
  /auth/password-update:
    post:
      tags: [Auth]
      summary: Update the caller's own password
      description: Requires a bearer token (either a normal session, or the session established right after following a password-reset email link).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties: { newPassword: { type: string, minLength: 8 } }
      responses:
        "204": { description: Password updated }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/exchange-code:
    post:
      tags: [Auth]
      summary: Exchange a PKCE code for a session
      description: |
        Only exercised if the Supabase project is configured for the PKCE
        flow (a `?code=` query param on the email link). This project's
        current config uses the *implicit* flow instead (tokens arrive
        directly in the URL fragment), which `app/auth/callback/page.tsx`
        handles client-side without calling this endpoint -- it's kept for
        if that configuration ever changes.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties: { code: { type: string } }
      responses:
        "200":
          description: Session established
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401":
          description: Invalid or expired code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/qr-login:
    post:
      tags: [Auth]
      summary: Sign in via a Master Tailor's QR code
      description: |
        Public -- the token comes from the `?token=` query param on the
        scanned QR's URL (`${WEB_APP_URL}/qr-login?token=...`). Internally
        looks up the token's hash against `qr_login_tokens`, then mints a
        real session via Supabase's admin `generateLink` + `verifyOtp`
        (the magic link is never emailed, just redeemed server-side
        immediately) -- see the README's Flow Map. Only ever issued for
        `master_tailor` accounts; deactivating an account or regenerating
        its QR immediately invalidates the old token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties: { token: { type: string } }
      responses:
        "200":
          description: Session established
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401":
          description: Invalid, expired, or revoked QR code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase-issued access token, obtained from `POST /auth/login` (or `/refresh`, `/exchange-code`). Verified independently on every request -- see `common/middleware/auth.middleware.ts`.
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      description: The one shape every error response takes, produced by common/middleware/error.middleware.ts regardless of which layer threw.
      properties:
        error: { type: string }
        code: { type: string, example: FORBIDDEN }
        details: {}
    Role:
      type: string
      enum: [owner_manager, designer, master_tailor, accountant]
...
    Profile:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    AuthSession:
      type: object
      description: Returned by login/refresh/exchange-code. Tokens only ever appear in this response body -- the API never sets a cookie itself.
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresAt: { type: integer, description: Unix seconds }
        profile: { $ref: "#/components/schemas/Profile" }
...