# src/modules/auth/api/auth.routes.ts
import { Router } from "express";
import { requireAuth } from "../../../common/middleware/auth.middleware";
import { asyncHandler } from "../../../common/http/async-handler";
import { validateBody } from "../../../common/http/validate.middleware";
import { AuthService } from "../application/auth.service";
import { DrizzleAuthRepository } from "../infrastructure/drizzle-auth.repository";
import { bootstrapDtoSchema } from "./dto/bootstrap.dto";
import { loginDtoSchema } from "./dto/login.dto";
import { refreshDtoSchema } from "./dto/refresh.dto";
import { passwordResetRequestDtoSchema } from "./dto/password-reset-request.dto";
import { passwordUpdateDtoSchema } from "./dto/password-update.dto";
import { exchangeCodeDtoSchema } from "./dto/exchange-code.dto";
import { qrLoginDtoSchema } from "./dto/qr-login.dto";
import { authProvider } from "../../../common/auth/supabase-auth-provider";

/** Composition root for the Auth module -- wires the concrete (Infrastructure) adapter into the Application service. */
const authService = new AuthService(new DrizzleAuthRepository(authProvider));

export const authRouter = Router();

authRouter.get(
  "/me",
  requireAuth,
  asyncHandler(async (req, res) => {
    res.json({ profile: req.profile });
  }),
);

authRouter.get(
  "/bootstrap-status",
  asyncHandler(async (_req, res) => {
    res.json(await authService.getBootstrapStatus());
  }),
);

authRouter.post(
  "/bootstrap",
  validateBody(bootstrapDtoSchema),
  asyncHandler(async (req, res) => {
    await authService.bootstrap(req.body);
    res.status(201).json({ ok: true });
  }),
);

authRouter.post(
  "/login",
  validateBody(loginDtoSchema),
  asyncHandler(async (req, res) => {
    res.json(await authService.login(req.body));
  }),
);

authRouter.post(
  "/refresh",
  validateBody(refreshDtoSchema),
  asyncHandler(async (req, res) => {
    res.json(await authService.refresh(req.body));
  }),
);

authRouter.post(
  "/logout",
  requireAuth,
  asyncHandler(async (req, res) => {
    await authService.logout(req.accessToken!);
    res.status(204).send();
  }),
);

authRouter.post(
  "/password-reset-request",
  validateBody(passwordResetRequestDtoSchema),
  asyncHandler(async (req, res) => {
    await authService.requestPasswordReset(req.body);
    // Always 204, whether or not the email matches an account -- never confirms existence.
    res.status(204).send();
  }),
);

authRouter.post(
  "/password-update",
  requireAuth,
  validateBody(passwordUpdateDtoSchema),
  asyncHandler(async (req, res) => {
    await authService.updatePassword(req.authUserId!, req.body);
    res.status(204).send();
  }),
);

authRouter.post(
  "/exchange-code",
  validateBody(exchangeCodeDtoSchema),
  asyncHandler(async (req, res) => {
    res.json(await authService.exchangeCode(req.body));
  }),
);

authRouter.post(
  "/qr-login",
  validateBody(qrLoginDtoSchema),
  asyncHandler(async (req, res) => {
    res.json(await authService.qrLogin(req.body));
  }),
);